+++ R/utils.R
@@ -170,10 +170,21 @@ getVarRand <- function(x) {
       sum(sapply(
         vals$vc[not.obs.terms],
         function(Sigma) {
-          Z <- vals$X[, rownames(Sigma), drop = FALSE]
+          rn <- rownames(Sigma)
+
+          if (!is.null(rn)) {
+            valid <- rownames(Sigma) %in% colnames(vals$X)
+            if (!all(valid)) {
+              rn <- rn[valid]
+              Sigma <- Sigma[valid, valid]
+            }
+          }
+
+          Z <- vals$X[, rn, drop = FALSE]
           Z.m <- Z %*% Sigma
           return(sum(diag(crossprod(Z.m, Z))) / stats::nobs(x))
         }))
+
     },
     error = function(x) { 0 },
     warning = function(x) { 0 },